:zap: refactor(utils): improve the route builder
Move the builder to `utils` directory and convert into a class.
