+++ src/utils/base-router/create.ts
@@ -1,7 +1,13 @@
 import { NextFunction, Request, Response, Router } from 'express'
-import { StoreService } from '../../../types'
-import HttpResponse from '../../../network/response'
+import { StoreService } from '../../types'
+import HttpResponse from '../../network/response'
 
+/**
+ * Generate an endpoint to create a resource.
+ *
+ * @param router - Instance of an express router.
+ * @param service - Service to connect with the database.
+ */
 const create = (router: Router, service: StoreService): void => {
   router.post('/', async (req: Request, res: Response, next: NextFunction) => {
     try {
